dynamic ssr page for todo

+++ pages/api/todos/[todoId].js
@@ -5,10 +5,20 @@ dbConnect();
 export default async function handler(req, res) {
   const { method, query } = req;
 
+  if (method === "GET") {
+    const todo = await getOneTodo(query);
+    return res.status(200).json({ todo });
+  }
+
   //   delete todo
-  if (method === "DELETE") {
+  else if (method === "DELETE") {
     await Todo.findByIdAndDelete(query.todoId);
     const todos = await Todo.find({});
     return res.status(200).json({ todos, message: "succeeded" });
   }
+}
+
+export async function getOneTodo(query) {
+  const todo = await Todo.findById(query.todoId);
+  return todo;
 }
